refactor(API Decoupage Administratif): basuler la logique de transformation en dict dans le fichier API
