+++ common/api/decoupage_administratif.py
@@ -25,3 +25,39 @@ def fetch_communes_from_epci(epci):
     response = requests.get(f"{DECOUPAGE_ADMINISTRATIF_API_URL}/epcis/{epci}/communes?fields=code", timeout=5)
     response.raise_for_status()
     return response.json()
+
+
+def map_communes_infos():
+    """
+    Create a dict that maps cities with their name/department/region/epci
+    Input: a list of communes. {"nom":"L'Abergement-Clémenciat","code":"01001","codeDepartement":"01","siren":"210100012","codeEpci":"200069193","codeRegion":"84","codesPostaux":["01400"],"population":859}
+    Output: a dict with the commune's insee code as pivot. {"01001": {"city": "L'Abergement-Clémenciat", "department": "01", "region": "84", "epci": "200069193"}}
+    """
+    commune_details = {}
+    try:
+        communes = fetch_communes()
+        for commune in communes:
+            commune_details[commune["code"]] = {}
+            commune_details[commune["code"]]["city"] = commune.get("nom", None)
+            commune_details[commune["code"]]["department"] = commune.get("codeDepartement", None)
+            commune_details[commune["code"]]["region"] = commune.get("codeRegion", None)
+            commune_details[commune["code"]]["epci"] = commune.get("codeEpci", None)
+    except requests.exceptions.HTTPError as e:
+        logger.info(e)
+    return commune_details
+
+
+def map_epcis_code_name():
+    """
+    Create a dict that maps epci with their name
+    Input: a list of epcis. {"nom":"CC Faucigny - Glières","code":"200000172","codesDepartements":["74"],"codesRegions":["84"],"population":28363}
+    Output: a dict with the epci's code as pivot. {"200000172": "CC Faucigny - Glières"}
+    """
+    epci_names = {}
+    try:
+        epcis = fetch_epcis()
+        for epci in epcis:
+            epci_names[epci["code"]] = epci["nom"]
+    except requests.exceptions.HTTPError as e:
+        logger.info(e)
+    return epci_names
